requirements.txt
amqp==5.0.2
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.3.0
bleach==3.2.1
boto3==1.16.55
botocore==1.19.55
cached-property==1.5.2
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
Django==3.1.5
django-celery-results==2.0.0
django-extensions==3.1.0
djangorestframework==3.12.2
docker==4.4.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
environs==9.3.0
gunicorn==20.0.4
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kombu==5.0.2
MarkupSafe==1.1.1
marshmallow==3.10.0
mistune==0.8.4
mongoengine==0.22.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
packaging==20.8
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.10
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pymongo==3.11.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
pyzmq==21.0.0
qtconsole==5.0.1
QtPy==1.9.0
redis==3.5.3
requests==2.25.1
s3transfer==0.3.4
Send2Trash==1.5.0
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
texttable==1.6.3
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
